Thread minimum RV through
